ai_engine/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0
openai==1.3.0
langchain==0.1.0
langchain-openai==0.0.2
langchain-core==0.1.23
psycopg2-binary==2.9.9
httpx==0.25.2
PyJWT==2.8.0

